Feat [Golang] Package documentation & implement main command
- [+] feat(exporter/session.go): add package documentation and types for chat session data extraction and conversion
- [+] feat(main.go): add main function and user input handling for JSON file path and output format option
- [+] feat(main.go): implement processing of CSV output option and format options
- [+] feat(main.go): implement processing of Hugging Face dataset output option
- [+] feat(main.go): add function to prompt for user input and save output to file
